refactor(telegram): extract s3 outbound lifecycle clusters into mixin modules (epic #78791)
